Kod presunut z Program.cs na Lesson5.cs. Funkce Main ponechana v Progam.cs a pri exekuci programu si bere veskere funkce a predpisy z Lesson5.cs. Upraveny urcite nazvy funkci dle predpisu HW.  Toto je posledni krko pred mergem s branch Homework/Lecture5
